backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==42.0.0

# GCP
google-cloud-storage==2.14.0
google-cloud-logging==3.9.0
google-cloud-secret-manager==2.18.0

# Redis
redis==5.0.1
hiredis==2.3.2

# AI/LLM
anthropic==0.18.1
openai==1.12.0

# Utilities
python-dateutil==2.8.2
numpy==1.26.3
pandas==2.1.4
pydantic-core==2.14.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
